Prevent overlapping GBIF imports (#317)
Previously, if you double-clicked the submit button in the admin UI's
GBIF import form, the server would run both imports at the same time.
This could result in duplicate data being imported.

Add a locking mechanism so the second attempt will fast-fail before
making any changes.

The locking mechanism is implemented as a general-purpose service since
this is unlikely to be the last time we'll need application-wide locking.

+++ src/main/kotlin/com/terraformation/backend/species/db/GbifImporter.kt
@@ -3,6 +3,9 @@ package com.terraformation.backend.species.db
 import com.terraformation.backend.config.TerrawareServerConfig
 import com.terraformation.backend.customer.model.requirePermissions
 import com.terraformation.backend.db.GbifTaxonId
+import com.terraformation.backend.db.LockService
+import com.terraformation.backend.db.LockType
+import com.terraformation.backend.db.OperationInProgressException
 import com.terraformation.backend.db.tables.records.GbifDistributionsRecord
 import com.terraformation.backend.db.tables.records.GbifNameWordsRecord
 import com.terraformation.backend.db.tables.records.GbifTaxaRecord
@@ -62,6 +65,7 @@ class GbifImporter(
     private val config: TerrawareServerConfig,
     private val dslContext: DSLContext,
     private val fileStore: FileStore,
+    private val lockService: LockService,
 ) {
   companion object {
     /** Number of rows to insert in a single SQL statement. */
@@ -96,6 +100,10 @@ class GbifImporter(
 
     try {
       dslContext.transaction { _ ->
+        if (!lockService.tryExclusiveTransactional(LockType.GBIF_IMPORT)) {
+          throw OperationInProgressException("Another import is currently in progress.")
+        }
+
         deleteAll()
 
         val taxonIds = openFile(TAXON_FILENAME).use { inputStream -> importTaxa(inputStream) }
